# 01-monolith/xcommerce-monolithic-master/final/openapi/MEGSI.openapi.yaml
openapi: 3.1.0
info:
  title: MEGSI
  description: ''
  version: 1.0.0
...
paths:
...
  /rest/order/list:
    get:
      summary: List
      deprecated: false
      description: ''
      tags: []
      parameters:
        - name: Authorization
          in: header
          description: ''
          required: true
          example: Bearer {{token}}
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: {}
            examples: {}
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties: {}
          headers: {}
      security: []
  /rest/order/checkout:
    get:
      summary: Checkout
      deprecated: false
      description: ''
      tags: []
      parameters:
        - name: Authorization
          in: header
          description: ''
          required: true
          example: Bearer {{token}}
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: {}
            examples: {}
      responses:
        '200':
          description: ''
          content:
            application/octet-stream:
              schema:
                type: object
                properties: {}
          headers: {}
      security: []
  /rest/order/backOffice/list:
    get:
      summary: Back Office List
      deprecated: false
      description: ''
      tags: []
      parameters:
        - name: customerId
          in: query
          description: ''
          required: false
          example: '0'
          schema:
            type: string
        - name: Authorization
          in: header
          description: ''
          required: true
          example: Bearer {{token}}
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: {}
            examples: {}
      responses:
        '200':
          description: ''
          content:
            application/octet-stream:
              schema:
                type: object
                properties: {}
          headers: {}
      security: []
...
components:
...
  securitySchemes: {}
servers: []
security: []